refactor: estrai domain/schema_diff.py da cli/inspect/
schema_diff_payload in domain/schema_diff.py, comando Typer in cli/.
Aggiorna MCP a importare da domain/.
Azzera dipendenza MCP→cli/ per schema_diff_ops.
